Move the DMF parser to the client
