--- OpenDreamShared/Interface/InterfaceDescriptor.cs
@@ -1,25 +1,0 @@
-﻿using System.Collections.Generic;
-using Robust.Shared.Analyzers;
-using Robust.Shared.Serialization.Manager.Attributes;
-
-namespace OpenDreamShared.Interface {
-    public sealed class InterfaceDescriptor {
-        public List<WindowDescriptor> WindowDescriptors;
-        public List<MacroSetDescriptor> MacroSetDescriptors;
-        public Dictionary<string, MenuDescriptor> MenuDescriptors;
-
-        public InterfaceDescriptor(List<WindowDescriptor> windowDescriptors, List<MacroSetDescriptor> macroSetDescriptors, Dictionary<string, MenuDescriptor> menuDescriptors) {
-            WindowDescriptors = windowDescriptors;
-            MacroSetDescriptors = macroSetDescriptors;
-            MenuDescriptors = menuDescriptors;
-        }
-    }
-
-    [Virtual, ImplicitDataDefinitionForInheritors]
-    public class ElementDescriptor {
-        [DataField("type")]
-        public string Type;
-        [DataField("name")]
-        public string Name;
-    }
-}
